requirements.txt
pytest
requests
allure-pytest
pytest-order
pytest-dependency
PyYAML